Minor fix to main file
originally created this website in a different folder, once I migrated it into the new github folder, it was throwing an error message on my dog list and owner list page, stating that cryptography package was required. I already had cryptography, so I was a little confused, so I did 'import cryptography' at the top of the page and it fixed the issue.